Primera/requirements.txt
facenet-pytorch
torch
Pillow
numpy
psutil
matplotlib
PyYAML